+Client Side User Verification Added To Utils
verifies user by token on every page.

NEXT: Resrict access based of user to those pages.
