--- client/src/utils/routeController.js
@@ -1,8 +1,0 @@
-
-
-
-module.exports = {
-
-    
-
-}
